Envio em massa para email, e solo para telegram

+++ src/components/pages/GruposUser.jsx
@@ -4,6 +4,8 @@ import axios from "axios";
 import {BsPencil, BsFillTrashFill} from "react-icons/bs"
 import React, { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
+import img from "../../img/blog_mini_4.jpg";
+import Swal from 'sweetalert2';
 
 function GruposUser() {
     const navegação = useNavigate();
@@ -23,43 +25,32 @@ function GruposUser() {
       coordenador: { id: null }
     });
     const [modoEdicao, setModoEdicao] = useState(false);
-    const [mensagemErro, setMensagemErro] = useState("");
 
     useEffect(() => {
       const accessToken = localStorage.getItem("accessToken");
-    
       if (!accessToken) {
         navegação("/Login");
       } 
-
     }, [navegação]);
 
     const cadastrarGrupo = () => {
       novoGrupo.periodo = { id: periodoId };
       novoGrupo.coordenador = { id: coordenadorId };
-      axios
-        .post("http://127.0.0.1:5000/api/grupo", novoGrupo)
+      axios.post("http://127.0.0.1:5000/api/grupo", novoGrupo)
         .then((response) => {
-          setModoEdicao(false);
           window.location.reload();
         })
         .catch((error) => {
-          if (error.response && error.response.data && error.response.data.message) {
-            setMensagemErro(error.response.data.message);
-          } else {
             console.log(error);
-          }
-        });
+          });
     };
     useEffect(() => {
-      axios
-        .get("http://127.0.0.1:5000/api/periodo")
+      axios.get("http://127.0.0.1:5000/api/periodo")
         .then((response) => setPeriodos(response.data))
         .catch((error) => console.log(error));
     }, []);
     useEffect(() => {
-      axios
-        .get("http://127.0.0.1:5000/api/coordenador")
+      axios.get("http://127.0.0.1:5000/api/coordenador")
         .then((response) => setCoordenadores(response.data))
         .catch((error) => console.log(error));
     }, []);
@@ -71,14 +62,36 @@ function GruposUser() {
       }, []);
 
       const removeGrupo = (id) => {
-        axios.delete(`http://127.0.0.1:5000/api/grupo/${id}`)
-          .then((response) => {
-            const AttListaGrupo = grupos.filter((grupo) => grupo.id !== id);
-            setGrupos(AttListaGrupo);
-          })
-          .catch((error) => {
-            console.log(error);
-          });
+        Swal.fire({
+          title: 'Você tem certeza?',
+          text: 'Se você apagar não tem mais volta!',
+          icon: 'Cuidado!',
+          showCancelButton: true,
+          confirmButtonColor: '#03A64A',
+          color: '#FFF',
+          background: 'rgb(32, 32, 36)',
+          cancelButtonColor: '#d33',
+          confirmButtonText: 'Sim, deletar!'
+        }).then((result) => {
+          if (result.isConfirmed) {
+            axios.delete(`http://127.0.0.1:5000/api/grupo/${id}`)
+              .then((response) => {
+                const AttListaGrupo = grupos.filter((grupo) => grupo.id !== id);
+                Swal.fire({
+                  title: 'Deletado!',
+                  text: 'O aluno foi apagado com sucesso',
+                  icon: 'success',
+                  confirmButtonColor: '#03A64A',
+                  color: '#FFF',
+                  background: 'rgb(32, 32, 36)',
+                });
+                setGrupos(AttListaGrupo);
+              })
+              .catch((error) => {
+                console.log(error);
+              });
+          }
+        });
       };
 
       const editGrupo = (id) => {
@@ -114,9 +127,8 @@ function GruposUser() {
       grupo.titulo && grupo.titulo.toLowerCase().includes(filtroGrupo.toLowerCase())
       );
 
-      const adicionarGrupo = () => {
-        setModoEdicao(true);
-        setMensagemErro("");
+      const adicionarGrupo = (e) => {
+        setModoEdicao(!modoEdicao);
       };
 
     return(
@@ -125,20 +137,18 @@ function GruposUser() {
           <main className={styles.conteudo}>
             <h1>Grupos</h1>
               {!editGrupoId && (
-                <form className={styles.form_filtro}>
-                <input type="text" placeholder="Pesquisar" value={filtroGrupo} onChange={(e) => setFiltroGrupo(e.target.value)}/>
-                <button text="Adicionar" onClick={adicionarGrupo}>Enviar</button>
-              </form>
+                <form className={styles.form_filtro} style={{ display: modoEdicao ? 'none' : 'flex' }}>
+                  <input type="text" placeholder="Pesquisar" value={filtroGrupo} onChange={(e) => setFiltroGrupo(e.target.value)}/> 
+                  <button type="button" onClick={adicionarGrupo}>Adicionar</button>
+                </form>
               )}
               {editGrupoId ? (
-                <div className={styles.editForm}>
-                  <form  onSubmit={saveEditGrupo}>
+                <div className={styles.EditForm_Alunos}>
+                  <form onSubmit={saveEditGrupo} className={styles.FormAlunos}>
                     <label>Título:</label>
-                    <input type="text" value={editGruposDados.titulo} onChange={(e) => setEditGruposDados({ ...editGruposDados, titulo: e.target.value })} required />
+                    <input type="text" placeholder="Digite o titulo do grupo" value={editGruposDados.titulo} onChange={(e) => setEditGruposDados({ ...editGruposDados, titulo: e.target.value })} required />
                     <label>Link:</label>
-                    <input type="text" value={editGruposDados.link} onChange={(e) => setEditGruposDados({ ...editGruposDados, link: e.target.value })} required />
-                    <label>Mensagem:</label>
-                    <textarea type="text" value={editGruposDados.mensagem} onChange={(e) => setEditGruposDados({ ...editGruposDados, mensagem: e.target.value })} required />
+                    <input type="text" placeholder="Digite o link do grupo" value={editGruposDados.link} onChange={(e) => setEditGruposDados({ ...editGruposDados, link: e.target.value })} required />
                     <label>Período:</label>
                     <select value={editGruposDados.periodo.id} onChange={(e) => setEditGruposDados({...editGruposDados,periodo: { id: e.target.value }})}>
                     <option value="">Selecione um periodo</option>
@@ -157,32 +167,36 @@ function GruposUser() {
                         </option>
                       ))}
                     </select>
+                    <label>Mensagem:</label>
+                    <textarea type="text" placeholder="Digite sua mensagem..." value={editGruposDados.mensagem} onChange={(e) => setEditGruposDados({ ...editGruposDados, mensagem: e.target.value })} required />
                   <button type="submit">Salvar</button>
                   </form>
                 </div>
               ) : (
-              <section className={styles.section_tabela}>
-                <table className={`${styles.table} ${modoEdicao ? styles.hidden : ""}`} style={{ display: modoEdicao ? "none" : "table" }}>
+              <section className={styles.section_tabela} style={{ display: modoEdicao ? 'none' : 'flex' }}>
+                <table className={styles.table}>
                   <thead>
                     <tr>
-                      <th>Id</th>
+                      <th>Grupo</th>
                       <th>Titulo</th>
                       <th>Link</th>
                       <th>Mensagem</th>
                       <th>Periodo</th>
                       <th>Coordenador</th>
+                      <th></th>
                     </tr>
                   </thead>
                   <tbody>
                     {filtro_Grupo.map((grupo) => {
                         return (
                           <tr key={grupo.id}>
-                            <td>{grupo.id}</td>
+                            <td><img src={img} alt="imagem tabela"/></td>
                             <td>{grupo.titulo}</td>
                             <td>{grupo.link}</td>
                             <td>{grupo.mensagem}</td>
                             <td>{grupo.periodo ? grupo.periodo.semestrereferencia : ""}</td>
                             <td>{grupo.coordenador ? grupo.coordenador.nome : ""}</td>
+                            <td>
                             <div className={styles.icones}>
                               <button onClick={() => editGrupo(grupo.id)} className={styles.icone1}>
                                 <BsPencil />
@@ -190,78 +204,45 @@ function GruposUser() {
                               <button onClick={() => removeGrupo(grupo.id)} className={styles.icone2}>
                                 <BsFillTrashFill />
                               </button>
-                            </div>
+                          </div></td>
                           </tr>
                         );
                       })}
                   </tbody>
                 </table>
-                {modoEdicao ? (
-                  <div className={styles.editForm}>
-                    <form onSubmit={cadastrarGrupo}>
-                    <div className={styles.formGroup}>
-                      <label>Título:</label>
-                      <input
-                        type="text"
-                        value={novoGrupo.titulo}
-                        onChange={(e) => setNovoGrupo({ ...novoGrupo, titulo: e.target.value })}
-                        required
-                      />
-                    </div>
-                    <div className={styles.formGroup}>
-                      <label>Link:</label>
-                      <input
-                        type="text"
-                        value={novoGrupo.link}
-                        onChange={(e) => setNovoGrupo({ ...novoGrupo, link: e.target.value })}
-                        required
-                      />
-                    </div>
-                    <div className={styles.formGroup}>
-                      <label>Mensagem:</label>
-                      <textarea
-                        type="text"
-                        value={novoGrupo.mensagem}
-                        onChange={(e) => setNovoGrupo({ ...novoGrupo, mensagem: e.target.value })}
-                        required
-                      />
-                    </div>
-                    <div className={styles.formGroup}>
-                      <label>Período:</label>
-                      <select
-                        value={periodoId}
-                        onChange={(e) => setPeriodoId(e.target.value)}
-                        required
-                      >
-                        <option value="">Selecione um periodo</option>
-                        {periodos.map((periodo) => (
-                          <option key={periodo.id} value={periodo.id}>
-                            {periodo.semestrereferencia}
-                          </option>
-                        ))}
-                      </select>
-                    </div>
-                    <div className={styles.formGroup}>
-                      <label>Coordenador:</label>
-                      <select
-                        value={coordenadorId}
-                        onChange={(e) => setCoordenadorId(e.target.value)}
-                        required
-                      >
-                        <option value="">Selecione um coordenador</option>
-                        {coordenadores.map((coordenador) => (
-                          <option key={coordenador.id} value={coordenador.id}>
-                            {coordenador.nome}
-                          </option>
-                        ))}
-                      </select>
-                    </div>
-                    <button type="submit">Cadastrar</button>
-                    {mensagemErro && <p>{mensagemErro}</p>}
-                    </form>
-                  </div>
-                ) : null}
               </section>
+            )}
+            {modoEdicao && (
+              <div className={styles.addGrupo}>
+                <form onSubmit={cadastrarGrupo}>
+                <h2>Adicionar Grupos</h2>
+                  <label>Título:</label>
+                  <input type="text" placeholder="Digite o titulo do grupo" value={novoGrupo.titulo} onChange={(e) => setNovoGrupo({ ...novoGrupo, titulo: e.target.value })} required />
+                  <label>Link:</label>
+                  <input type="text" placeholder="Digite o link do grupo" value={novoGrupo.link} onChange={(e) => setNovoGrupo({ ...novoGrupo, link: e.target.value })} required />
+                  <label>Período:</label>
+                  <select value={periodoId} onChange={(e) => setPeriodoId(e.target.value)} required >
+                    <option value="">Selecione um periodo</option>
+                    {periodos.map((periodo) => (
+                      <option key={periodo.id} value={periodo.id}>
+                        {periodo.semestrereferencia}
+                      </option>
+                    ))}
+                  </select>
+                  <label>Coordenador:</label>
+                  <select value={coordenadorId} onChange={(e) => setCoordenadorId(e.target.value)} required >
+                    <option value="">Selecione um coordenador</option>
+                    {coordenadores.map((coordenador) => (
+                      <option key={coordenador.id} value={coordenador.id}>
+                        {coordenador.nome}
+                      </option>
+                    ))}
+                  </select>
+                  <label>Mensagem:</label>
+                  <textarea type="text" placeholder="Digite sua mensagem..." value={novoGrupo.mensagem} onChange={(e) => setNovoGrupo({ ...novoGrupo, mensagem: e.target.value })} required />
+                <button type="submit">Cadastrar</button>
+                </form>
+              </div>
             )}
             </main>
       </>
